requirements.txt
google-genai
python-dotenv
psycopg2-binary
python-telegram-bot
numpy
pgvector
